Add new option for LineUI
